build: building with bootstrapped LLVM
Infrastructure and changes for building CPP code with downloaded and
built LLVM. The LLVM is built in bootstrap mode i.e. first minimal
Clang toolchain is built with system compiler in the next step it is
used to build more optimized compiler and toolchain.  Warning:
Building internal LLVM for the first time takes ~2h

LLVM is not built by default, in order to use it one need pass
'--clang internal' parameter to the 'build.py' tool.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
Message-Id: <20190514134345.14825-1-michal@vectorized.io>