bytes: Add bytes_ostream, a fragmented buffer
The bytes_ostream class represents a buffer where the size is not
known up-front, and is therefore fragmented with no copying of
contents.

Consumers of this class write bytes to it. It supports the notion of
placeholders, bytes for which the consumer obtains a pointer to fill
in later.

A bytes_ostream can be linearized, where its contents become a
single, sequential buffer, meant to be used for testing.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>