utils: Add concept for a fragment range
Large, contiguous allocations are a problem for systems due to
fragmentation. They can also add latency, since they can require the
memory allocator to work harder to allocate that memory (e.g., it may
require the seastar allocator to make external calls to the memory
reclaimers). Indeed, seastar emits warnings about large memory
allocations.

This patch adds a concept for data structures which are fragmented,
and can thus be read as a range of fragments.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>