redpanda/kafka: Add basic Kafka API server
FIXME: Split this into smaller patches.

Introduce the Kafka API server, which listens for and accepts client
connections. It ensures that, for a connection, responses are sent in
the same order the corresponding requests were received in. It also
ensures an upper limit on the amount of memory the server will
consume, at which point it will not read more requests from the
socket, causing backpressure.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>