utils: Add remote marker type
The remote<> marker type wraps a graph of objects that can be moved
across shards, but which the programmer ensures will only be deleted
on the origin shard.

This is useful for request-response scenarios, where the remote<> type
is passed to a shard, which then moves it back to the origin one
together with some response type.

The advantage is that it saves one inter-shard message.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>