redpanda/kafka: Add smp service group to avoid overwhelming target CPUs
When sending requets to a remote shard, concurrency can be controlled
by a semaphore. However, many shards (up to smp::count - 1) can send
many requests to the same remote shard, enough to overwhelm it.

In the context of the kafka API, each shard handles a set of
connections, on which they can receive requests that need to be sent
to other shards. This patch helps the remote shards apply backpresurre
to the shards processing the connections.