rpk: fallback to ntpstat even if timedatectl is avaialable
In order to make it possible to run `rpk check` without errors in
containers running on host with systemd installed. Prevented NtpSynced
function from returning an error when `timedatectl` does not exit
cleanly. Falling back to `ntpstat` instead. The `timedatectl` is present
in containers running at host with systemd but when executed within a
container it returns an error.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
Message-Id: <20190731100215.25014-1-michal@vectorized.io>