storage: Add log segment reader
The log segment reader uses the parser to read record batch off disk.
It uses a consumer which skips to the first offset while we don't
implement an offset map.

It also contains logic to determine when a read is done, or example
when we reach the request max bytes.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>