rpk/checkers: implemented checkers for disk scheduler and nomerges
In order to check if disk scheduler is properly set before starting
redpanda, implemented scheduler and nomerges settings checkers. Scheduler
checker checks if disk scheduler is set to either `noop` or `none`,
the nomerges checker checks if disk nomerges setting is set to 2
(disable merging). The checkers will be used in disk tuners as well.

Signed-off-by: Michal Maslanka <michal@vectorized.io>