tools:rpcgen.py: add rpc code generator based on jinja2 templates
This is a lightweight tool to generate a header file from a json input.
The output of this codegenerator is a fully working and compilable service
the user can simply extend with very little effort.

Note that the code generated and is both very small and _strictly_ optional
a user is free to manually extend and implement the service.h interface
and to generate a client herself.

Typical input:

{
    "namespace": "cycling",
    "service_name": "team_movistar",
    "includes": [
        "rpc/test/rpc_gen_types.h"
    ],
    "methods": [
        {
            "name": "canyon",
            "input_type": "ultimate_cf_slx",
            "output_type": "nairo_quintana"
        },
        {
            "name": "ibis_hakka",
            "input_type": "san_francisco",
            "output_type": "mount_tamalpais"
        }

    ]
}

will generate 2 classes. `team_movistar_service` and
`team_movistar_service::client` respectively