rpk: executor interface definition
Definition on an `Executor` interface. This interface will be use by the
tuners to execute the changes in operating system. Currently the tuners
interact with:
- filesystem
- sysctl
- ethtool

To perform changes needed to optimize system performance, the executors
will intercept this calls and execute actions which will be wrapped in
commands. Each command will have a logic to either execute an action
directly or render a script that will execute the same action when
launched.

Signed-off-by: Michal Maslanka <michal@vectorized.io>