storage: Add log segment representation
A log is composed of one or more log segments. The log_segment class
models an individual segment, and contains function to obtains read and
write streams, as well as management functions for the underlying data
file.

The log segment maintains an exclusive upper bound for the offsets written
in it. That upper bound is derived from:

1) The base offset of the next segment; or
2) The newest write, if that segment is the active one.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>