storage: Add log segment parser
The parser is able to read record batches from a log segment. It reads
the file using an input_stream, and handles data types crossing over
two or more temporary_buffers returned by the input_stream.

The parser is extensible via the consumer class, which can be used to
control the parsing and implement higher-level readers.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>