model: Add record_batch_reader
The record_batch_reader is a high-level reader interface for reading
record_batches. It's designed to allow a slice of batches to be read
from an underlying source, and then consumed from an upper layer
without any deferrals.

That underlying source can be a log segment or an in-memory
representation of a segment.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>