storage: Add the offset tracker class
Log segment readers cannot read past the latest committed offset. A
committed offset is specified by the replication protocol (Raft, in
our case). Using this class, the writers can communicate to the
readers what the latest committed offset is.

We also track the latest dirty offset, with the invariant that the
latter is always bigger or equal to the committed offset.

This will later be used by the log class, which we forward-declare
here.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>