utils: Add copy_range future combinator
The copy_range combinator, inspired by boost::copy_range<>, transforms
an input range, sequentially applying to each element the specified
asynchronous function, and building a result container with each
returned element.

This patch is extracted from a pending patch set to seastar. If that
patch series makes it in, this file will be delete. Otherwise, the
unit tests will be added here, as well as the parallel version.

Signed-off-by: Duarte Nunes <duarte@vectorized.io>
Message-Id: <20190914004047.8890-1-duarte@vectorized.io>