e2e_tests: support for running test against custom cluster
Implemented support for running the tests against a cluster of Kafka
nodes specified with an `--address` flag. Using ClientFactory class
to abstract the creation of Kafka Consumer, Producer and Admin type
clients. Thanks to the factory pattern usage the test logic does not
depend on that whether the clients connect to remote/local broker,
using address provided via command line parameters or to sandbox
nodes.

Signed-off-by: Michal Maslanka <michal@vectorized.io>