docs/rfc: add rfc for extending record_batch_header
The detection whether or not the record is internal should actually
happen at the batch level. The reason is `record_batch` is our atomic unit for
reads and writes.

Thus, the next modifications follow:

1. add a `model::record_batch_type` _above_ the CRC line
2. de-virtualize `raft::entry` type

We want other subsystems `raft` and `cluster` in particular to take advantange
of our `storage::log` and related infrastructure to persist changes to disk.
It allows us to have one code base for persisting records and reading them from
disk.