Fix bash commands in README.md
The line continuations (\) in the commands weren't working because of
the inlined comments: if the \ is placed before the comment, it
has no effect because it isn't followed by a newline. If it's placed
after the comment, it's commented out and the next line is evaluated
on its own.

If we wanna keep the inline comments, the cleanest solution is to
surround them with backticks (`).

Note that this effectively evaluates them in a subshell, which adds
some overhead, but it's barely perceptible given that there aren't
many comments.

Signed-off-by: David Castillo <david@vectorized.io>
Message-Id: <20191009183515.27398-2-david@vectorized.io>