build:cmake: default -march=westmere
Fixes: #187

At the moment, use the research of the CPU arch that scylla is doing
since they have a much wider support of scylla and have 4 years of
experience with running on a wide variety of architectures.

We can at some point turn this into a paid feature, where we allow
the users to extract the maximum performance by specifying their
lowest common denominator CPU.

From looking at the wiki:
https://en.wikipedia.org/wiki/Westmere_(microarchitecture)#Westmere_CPUs
it looks like westmere is now 11 years old, so we can use that as a
measurement going forward w/ the CPU types that we can support.

since v1:

* fix user name

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20191010190932.3025-1-alex@vectorized.io>