kafka/request_reader: fixed error in reading nullable strings
The following rule describes NULLABLE_STRING encoding in Kafka protocol:
For non-null strings, first the length N is given as an INT16. Then N
bytes follow which are the UTF-8 encoding of the character sequence. A
null value is encoded with length of -1 and there are no following
bytes. request_reader when reading the NULLABLE_STRING read the string
length twice causing an error. Fixed the bug by using already read
length instead of reading next two bytes.

Signed-off-by: Michal Maslanka <michal@vectorized.io>