Port tools/verify_git_cfg.sh to python
Port the git config-checking script to python to make it easier to modify
and read.

Add a check to verify that the from field in [sendemail] has the correct
format, to avoid patch messages being omitted on the mailing list.

Adds a ValidationRule class, and related subclasses, to make it easy to
add new ones, as well as to save the exceptions, if any, to log them
when log = DEBUG.