build/cmake: adds downloading of llvm source code
When 3rd party external projects are installed on a system that already
has clang available (that is, where we don't build clang ourselves), in
order to build libcxx and libcxxabi, we need to obtain the LLVM source
code. This commit adds a new 3rdparty project for downloading LLVM
source code (but not building clang).