devprod/dockerfile: adds base image
Adds a Dockerfile for creating an image where base 3rd party OS (fedora)
packages needed to build redpanda are installed. We first disable update
repos in order to ensure that we re-building the image does not result
in installing newer versions of these dependencies.

The .dockerignore file ignores everything but the tools/install-deps.sh
script. This allows us to use Docker's caching mechanism and avoid
having to rebuild the image unnecessarily.