devprod/dockerfile: adds image for 3rd party projects
Adds Dockerfile for building and installing 3rd party external projects
to /usr/local. We need to take care of removing version references in
Seastar's config since the .so files we produce do not include
versioning information and CMake errors out complaining about this.
Since we have full control of the stack, this is fine but this should
ideally be solved (see #191).

The build depends on CMakeLists.txt, 3rdparty.cmake.in and
cmake/FindLibCxx.cmake, as well as the base image (clang). Whenever
there's a change in any of these, the image needs to be rebuilt.