devprod/gcb: adds GCB pipeline
The pipeline builds docker images for the toolchain, making use of the
GCR/Kaniko caching service. The pipeline defaults to using Clang to
compile and "Release" as the build type. These can be changed by making
use of GCB's substitutions (parametrization) mechanism.

A new section to the docs/hacking.md file is added that serves as a
quickstart guide on running the GCB pipeline locally.

fixes #177