devprod/dockerfile: move to monolithic builder image
The Dockerfiles we had were all referencing to an already-existing
docker image on the gcr.io registry (fedora:30 tag). This meant that the
GCB pipeline was using these images (fedora-30 tag) as opposed to using
the images that were built as part of the pipeline execution. In order
to use these images, we would need to tag them with a value that is
defined at runtime. This is currently not supported by GCB though:

  https://github.com/GoogleContainerTools/kaniko/issues/713

Thus, if we want to use kaniko and build RP using the builder image that
is created as part of the execution of the GCB pipeline, we need to have
a single Dockerfile. Thus, we move from having multiple Dockerfiles to a
single one.