devprod/dockerfile: copies iotune out of seastar build dir
Seastar's CMake config does not add any target defined in seastar/apps/*
to the list of targets that are to be installed when 'make install' is
executed for seastar. In our case, we need the iotune app to be
installed since it is included in the RP packages. Therefore, we extend
the Dockerfile so that it copies iotune out the build dir into
/usr/local/bin