rpk: cmd/iotune.go: Add --timeout flag
Unlike the other commands' timeouts, the iotune timeout is by
default the max allowed duration (which amounts to not having a
timeout) and is added to the value given to --duration.

For example, if the duration specified is of 30 minutes, and the
given timeout is 10 minutes, iotune will be stopped if it hasn't
completed after 40 minutes (30 + 10).