test_utils/fixture.h: improve fixtures to allow blocking ctor
Allow test fixtures to have blocking constructors and destructors.
The previous version was a hotfix on master, to make the build
pass the tests suites.

Usage:

class fixture {
    fixture() {
        _f = open_file_dma("foo").get0();
    }
    ~fixture() {
        _f.close().get();
    }
    file _f;
};

FIXTURE_TEST(file_ops, fixture) {
    _f.write_dma("").get();
}

This test will now _not_ leak any resources as we've
waited for all the fibers to finish

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20191025205914.19063-1-alex@vectorized.io>