raft/consensus_utils: add clone_entries() method with unit tests
Raft needs to notify users if it has committed the entries.
The api we expose is that we give the user a copy of the raft
entries comitted. Given that both the filesystem
and the notification callback are sinks of data, we clone the
readers in memory before hitting disk, so each sink can have
their own light weight copy (just tmp-bufs share())