devprod/gcb: adds packagecloud step
Adds a step to publish release packages to packagecloud by reading git
tag information. A TAG_NAME variable is queried and, if exists and
points to a 'release-*' commit, it publishes the generated packages to
packagecloud.

The documentation on how to trigger the gcb pipeline locally is also
modified accordingly.

fixes #236