cluster/controller: wired up the entries commited notification clb
The controller have to be aware of all the metadata changes in order to
update the metadata cache so that each node of the cluster is aware of
partition placement replicas and leaders. The information will be
passed by raft0 consensus via commited entries notification callback.
The callback will trigger the same path as used during recovery.

Signed-off-by: Michal Maslanka <michal@vectorized.io>