rename: cluster/simple_batch_builder -> storage/record_batch_builder
Rename batch builder helper function factoring out only the storage/utils
pieces into the storage hierarchy and leaving the RPC serialization
in the cluster utils as to not introduce a new dependency in storage.

We add a common interface and use inheritance for light level extensions
on storage::record_batch_builder. The plan is to extend this fluent
api to make unit testing easier in the future for all storage tests