storage/log.h: add a high level flush() method to log
Flushing is _not_ a trivial operation. It is more than just
calling flush() on the underlying file handle. It requires the
log to update the offset trackers for dirty & committed offsets.
It also requires the log to update the last appender's max
offset. It is not possible to flush the log correctly by
flushing the underlying writer.

This must be a first class method