k: add kafka api serde tool
this tool is intended to be used to perform various operations on the
serialization/deserialization of kafka requests.

currently it implements an echo mode in which a binary encoded request
is decoded from stdin using the redpanda decoder, then encoded and
output using the redpanda encoder. the current usage is to verify from a
gold source our encoder/decoder.

Signed-off-by: Noah Watkins <noah@vectorized.io>