cluster/controller: waiting for recovery to end
The partitions have to be fully recovered before redpanda can start
serving requests. Previously the recovery was done in the background.
That caused the race condition to happen when other services accessed
the partitions that weren't fully recovered. In order to prevent this to
happen implementing waiting for all partitions to be recovered before
finishing the future<> returned from controller::start method.

Signed-off-by: Michal Maslanka <michal@vectorized.io>