devprod/dockerfile: refactors Dockerfile to use build.py
Invokes build.py from the Dockerfile in order to avoid having divergent
logic between what devs do and what is done in CI.

A set of helper bash scripts in tools/ci/scripts correspond to bash
scripts that get invoked from the Dockerfile. The goal of having these
outside of the Dockerfile is to enhance readability of the Dockerfile,
and to make it clear when the installation of LLVM (or external
dependencies) logic changes, as opposed to something specific to Docker.
These scripts handle the assignment of value to the --clang flag of the
tools/build.py command, as well as running post-installation tasks.