devprod/gcb: modifications due to new builder image
The pipeline is updated in order to reflect the changes to the builder
image. Configuration and compilation are done by build.py, so these
steps are consolidated into a single one.

Scripts added to tools/ci/scripts/ are invoked from the GCB pipeline in
order to enahance readability of YAML file, as well as to keep the logic
separate.

fixes #229