model/metadata: made partition_metadata to hold broker_shards
In order to store the information about placement of each partition on
particular node cores in metadata_cache change the partition_metadata to
hold broker_shard instead of just an id. The Kafka Metadata API does not
require the shard to be present but it can just be skipped when encoding
Metadata API response. The shard placement information will be used to
recover state of a partition_allocator.

Signed-off-by: Michal Maslanka <michal@vectorized.io>