cluster/recovery-test: using only current node as replica
The controller recovery test triggered an error when partition-allocation
happened. On fast machines (like build node) the test failed because
raft leader election triggered allocator recovery. The allocator was
missing brokers that were generated while setint up the test so the test failed.
This issue did not occurred on slower nodes as the leadership changed
notification was not triggered.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
Message-Id: <20191115153038.75777-1-michal@vectorized.io>