r/consensus: unified logic of self vote handling
Raft implementation was able to elect a leader while not having any members.
In order to simplify the logic and make it correct moved handling of self
vote to the same path as other vote requests. When vote should be requested
from current node instead of exchanging RPC call the implementation calls the
one_vote method. Majority counts are updated to classical raft (n/2)+1.

Signed-off-by: Michal Maslanka <michal@vectorized.io>