infra: Add terraform plan to deploy a 1-node test env on AWS
Add a plan to deploy a one-node cluster on AWS. The default distro is Fedora 31,
which uses kernel version 5.3, and the default instance type is l3.large,
which comes with 2 vCPUs, 15.25 GiB, and a 475 NVMe SSD.

The only supported region is us-west-1.

When started, the packagecloud repo is set up and redpanda is installed
and started (via systemd).