treewide: fixed logger definitions
The way that the logger was previously defined prevented seastar
framework from setting their log levels. As the static local variables
are lazy initialized the logger were added to global loggers registry
after the log level was set. In order to prevent this to happen using
variables that are initialized at the application start.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
Message-Id: <20191119181845.23461-1-michal@vectorized.io>