rpk: write_file.go: Open file with O_SYNC instead of calling Sync()
Calling Sync() (fsync) is not allowed on certain files - such as those
in sysfs (/proc/sys/) -, in some VMs and distros (e.g. Amazon Linux).
This caused the tuners that wrote to /proc/sys to fail.
This is fixed by opening the file with O_SYNC instead of calling Sync.

Signed-off-by: David Castillo <david@vectorized.io>
Message-Id: <20191119224125.4293-1-david@vectorized.io>