raft/consensus_utils: unify in-memory batch consumption/sharing
We had 3 distinct ways to collect in memory record batches
and then performing actions on them.

With the addition of foreign_share_n() we now unify all buffer
sharing code and batch aggregation code into a single code path

In addition we expose 2 new utilities for buffer sharing
which just propagate the iobuf_foreign_share_n() method
as an interface.