vtools: add vectorized tools nestable CLI
Our tools directory is getting unwieldy. This
patch introduces a build system, a testing framework,
and a nestable command line tooling with a single
utility ported which is our git verifyer with proper packages

to test:

```
cd tools
tox package
./.tox/py37/bin/vtools verify-git --path .
```

as you can see, our vtools is an entry point command with
chainable commands, in this case verify-git which corresponds
to the path `vtools/git/commands:verify_git` python function.