configuration: added advertised RPC address property
It may be required to differ the listen address of RPC server from the
one that is published to other cluster members. In order to do that
introduced advertised RCP address property. When present the address
defined by this property is going to be used by other cluster members
to connect to node configured with it.

Signed-off-by: Michal Maslanka <michal@vectorized.io>