c/controller: initializing consensus nodes with assignments
When created the raft group has empty configuration. The configuration
is populated during bootstrap (recovery) however when group is created
for the first state it have to be initialized with the list of nodes
that comes from assignment process. Modified code to pass all replicas
to the core when group will be created and then leveraging the
metadata_cache to collect broker metadata required to initailize group
configuration.

Signed-off-by: Michal Maslanka <michal@vectorized.io>