k: add fetch request partition iterator
the fetch request handler needs to iterate over the partition read
requests in order. the on-wire representation is a multi-level data
structure. in order to avoid creating a new data representation that
would result in data copying or constructing a separate index over the
data, this adds a multi-level iterator over the raw request data.

Signed-off-by: Noah Watkins <noah@vectorized.io>