r/consensus: refactor voting into it's own state machine
Fixes subtle concurrency bug where in order to achieve convergence
groups must only lock for smaller periods of time and not the full
duration of the vote session.

This differs from the locking strategy for replicate() because
we must gurantee linearizability of writes - both to our log
and to remotes.