debian:packaging remove shlibs:depends since we have none
As per debian discussion:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=141730

"""
> dpkg-gencontrol: warning: unknown substitution variable ${shlibs:Depends}

This is a bug in your package then. If your package does not set up a
substvars file with shlibs:Depends in it, dpkg-gencontrol will have
nothing to substitute. Probably this package has no dependencies on
shared lbiraries and you should remove the ${shlibs:Depends} from
debian/control.
"""