iobuf: consume exactly one frag should move to next
when the size requested to iobuf iterator consume lands exactly on the
end of the active fragment control exits without advancing to the next
fragment. this means that things like is_finished may not report the
correct state.

upon returning from consume this patch adds a check if the iterator
needs to advance to the next fragment.

Signed-off-by: Noah Watkins <noah@vectorized.io>