cluster: add check for partition existence
a typical pattern is to (1) check locally which shard an ntp is located
on then (2) schedule a future to run on that shard which grabs a
reference to the ntp's partition from the partition manager.

there does not appear to be any synchronization that guarantees the
partition does not disappear from the partition manager. thus we add an
extra check on the target shard itself with the contains method added by
this patch.

Signed-off-by: Noah Watkins <noah@vectorized.io>