k: create api_versions api response type
previously the api_versions request handling implementation was handled
manually compared to other api endpoint request handling. this was
primarily because the request itself is not subject to the versioning
restrictions like other api and the request itself is an empty request.

this patch treats the api version request handling like other requests,
and importantly, builds a response type that we can use for
encoding/decoding like other api implementations.

the only exception is that request processing still skips the versioning
enforcement.

Signed-off-by: Noah Watkins <noah@vectorized.io>