devprod/gcb: adds gcloudignore to include .git/ folder
Adds a .gcloudignore file so that .git/ is not excluded from the
files mounted into the workspace by the gcb pipeline (it is ignored by
default). The .gcloudignore file includes Git's .gitignore, which is
useful when triggering the pipeline from a local machine; the downside
being that if the `.git/` folder is big, it will be part of the tarball
that it's placed in the gcs bucket.
Message-Id: <20191207061538.13125-1-ivo@vectorized.io>