k/test: add metadata request test
this uses the expected failures variant of the tests. this is useful
because it allows us to write specific use case tests that aren't
currently supported and which have associated asana tickets. once those
tickets have been completed the tests will actually report failure and
the expected annotations can be removed. this allows a nice decoupling
of concerns--writing tests and fixing problems/adding features.

Signed-off-by: Noah Watkins <noah@vectorized.io>