build: fixes clang-debug build
Changes libz config so that a shared library is not produced. This with
the goal of avoiding clang picking it up instead of the dev library
installed via OS package (zlib is itself a dependency of clang that gets
installed by tools/install-deps.sh). Also, we now make zlib a dependency
of libcxx (as opposed to have it pick up the system's).
Message-Id: <20191212022502.22582-1-ivo@vectorized.io>