devprod/docker: do not install deps on /usr/local
Modifies the builder image recipe (tools/ci/install-external.sh) so that
the installation of 3rd party dependencies is not done in /usr/local
anymore, and instead create the cmake build folder outside the working
directory, specifically on the /v folder. Since this is done at the time
the docker image is created, this folder persists (we take care of
removing the cmake cache and intermediary object files).

Subsequently, when the gcb pipline step instantiates the builder image
(tools/ci/scripts/build-and-test-cpp.sh), it configures the build so
that the /v folder is used as the cmake build directory. This
effectively allows to break the process in two without the need to move
folders or create symlinks; doing any of these break the build by not
persisting build-specific RPATHs.

The motivation behind this change is to mirror the exact same
environment of developers, in which all dependencies are located within
the build/ folder (build/llvm and build/$BUILD_TYPE/v_deps_*), and where
the paths for intermediary objects are not affected by moving or
symlinking the build/ folder.