Partition
The "Partition" type is an abstraction that is part of the public API.
It represents a single NTP tuple and its backing data store. The internals
of how data are stored, indexed, segmented, or retreived should not leak
beyond this type to higher-level types.

Still todo: partition index - an index of segment indices that locates
the subindex from the top-level read() call.