k/request: add produce request error checking
this adds error checking related to record batch correctness performed
by kafka but absent in redpanda. in particular, we now check if there
are any idempotent or transactional records (which rp does not support).
and we also verify that exactly one batch exists and that it is of the
expected format.

Signed-off-by: Noah Watkins <noah@vectorized.io>