s/log: fixed incorrect offset assignment
Right now all the storage offset tracker consistently tracks the offset
bounds in the way that lower bound is inclusive while upper bound is
exclusive. This approach is convenient and fast as most of the offset
management logic is realized through offset arithmetic. In order to keep
the offset tracking consistent it is required to used
model::record_batch::end_offset instead of last_offset. This patch fixes
incorrect offset tracking in storage layer.

Signed-off-by: Michal Maslanka <michal@vectorized.io>