storage: use clear semantics when opening segments
Simplifies the semantics such that we can use the routine to test if a
file is a segment, identify a segment that is successfully opened, and
identify situations where an error occurred when opening a segment.

The distintion between an error opening a segment and a file being or
not being a segment is important. We do not ever want to ignore a
segment file just because it could not have been opened. Doing this
would lead to horrible bugs like ignoring the latest segment. This
distinction was not particularly clear in the previous verison of the
code.

Signed-off-by: Noah Watkins <noah@vectorized.io>