model: add init phase to record batch reader consumer
adds a specialization for the consumer interface that allows a consumer
to specialize an initialization method that is called before the
consumer is given its first record.

this is useful for consumer that need to initialize things like the log
appender that may create a segment if one does not yet exist.

Signed-off-by: Noah Watkins <noah@vectorized.io>