rpc/connection: fixed closing server connections
Server connections should be closed before leaving the gate as in gate
the server is waiting for input. Connections are kept alive by the
client so in order to minimize latency the server has to keep polling
the underlying fd, via input_stream::read_exactly call. When there will
be enough bytes the RPC dispatch logic will continue. In order to
interrupt the wait output fd have to be closed.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
Message-Id: <20200119173833.74256-2-michal@vectorized.io>