storage: add indexing structure to segment through log_manager
Instead of recovering the last segment, this patch adds in structure
to pass in all the requirements to construct a segment with indexing.

Currently, indexing is not involved in the reads or writes, but
this is the first infrastructure patch that changes the interface
for the storage layer to accept an index, reader and writer
as the building block

Subsequent patches will proxy the calls from segment to underlying
handles going through indexing layer.

storage: fix unit test for indexing