s/log_set: add proper lower_bound semantics.
std::lower_bound, given an order list, it will return the iterator
to the next elemnt. That is greater than or equal to, but not less.

[ 1, 2, 4, 6, 6, 6, 7, 8]

Doing a lower bound on 4, yields the index pointing to element 6

[ 1, 2, 4, 6, 6, 6, 7, 8]
           ^
           lower bound

However, our ranges are buckets of inclusve ends which means that
when we do a lower_bound on dirty_offset() we need to check if
we are in range at all or if the segment is empty