s/disk_log_impl.cc: roll over segments that have data.
This is a bug that manifested when we would write _exactly_
log_segment worth of bytes. It would cause a segment roll for the
given term, but contain no data. Followed by a term roll over
it would mean that 2 terms contain the same offset which is a bug
in the storage. Ensure that we _only_ keep segments that have data
in memory before rolling over new segments.

This bug was discovered via `vassert()` in log_set.cc,
see previous commit on lower_bound.