storage: add indexing to storage for both reads and writes
This patch we proxy all write calls to the existing segment_offset_index.
The read calls expose a new method called `offset_data_stream(model::offset)`
which translates model::offset to physical offset nearest the offset point.

In addition to reads and writes, indexing eliminates the need for having
yet-another-piece-of-separate-state called `offset_tracker` which
was only intended for keeping track of dirty (append'ed objects)
and `committed` offsets.

As with all things filesystem, this change is resilient to crashes
and data corruption of indexes themselves which are _only_ flushed
to disk when `closed` at the moment. This will change
in a follow up patch.

During recovery/bootstrap, we attach to the `checksumming` consumers
and re-build the indexes for those indexes that we have not yet built
or were corrupted, in anyway.

Minor: there were 2 small tests that were disabled because they were
testing for obsolete behavior of the storage engine, but I wanted to
keep them to port them over once `truncation` is in.