s/parser.h: add state to stop parsing altogether
When we parse batches, there is no way to communicate to the driver
(parser.cc) to stop parsing corrupted state. Even if we detect
that a header is wrong (malformed input), we need to _not_ only
skip the batch, but actually stop parsing altogether.

Issuing calls to input_stream<char>::skip(<some large corrupted>)
value yields to an internal assertion inside seastar.

The correct behavior is to, stop the parser, reset the state and
stop iteration.

This patch exposes a secondary state to batch_consumers