utils/file_sanitizer: add named backtrace for debug
When the file sanitizer detects a bug, it is often hard to
understand *what* was the operation that trigger the concurreny bug.
Was it _close_ or read or maybe the input stream _read_bulk_
etc. This patch adds the name of the backtrace with some useful
parements printed at time of operation disaptch