storage/log_segment_reader: support truncating sealed segments
note: this file _must_ be open in `ro` mode only. Seastar uses dma files
with no shared buffer cache around them. When we use a writer w/ dma
at the same time as the reader, we need a way to synchronize filesytem
metadata. In order to prevent expensive synchronization primitives fsyncing
both *reads* and *writes* we open this file in ro mode and if raft requires
truncation, we open yet-another handle w/ rw mode just for the truncation
which gives us all the benefits of preventing x-file synchronization
This is fine, because truncation to sealed segments are supposed to be very
rare events. The hotpath of truncating the appender, is optimized.