storage/appender: enable file truncation with half-page readback
Assuming a no-cache (argument still holds with cache) file appender
which has written 3 full pages. Supporting a truncation for the middle
of the first page requires, us to readback the half page in order to
support continued DMA writes to this file handle without reopening
the file.

Example:

Page 1   Page 2  Page 3
[ 1 2 ] [ 3 4 ] [ 5 6 ]

At this point we have written, 3 full pages. Assume every entry
is exactly half a page.

log_segment_appender::truncate(3) requires us to read page 2, haf page
such that in memory we have

[ 3 _ ]
  ^

Where 3 is what we have in memory to be able to continue DMA writes
and

[ 3 _ ]
    ^

and `_` is our current pointer to the half page for the next append()
call