storage/log_reader: simplify interface for range queries using index
Now that we are reading 32KB per read to the continous parser,
we collapse one layer of abstraction and make the segment reads
use the index instead of keeping around state for tracking
and input_stream<char> object pointer.

Note that previous implementation leaked resources and never
closed the input stream in case of errors.