vtools/fmt: adds 'all' subcommand and --check flag
Adds a new 'vtools fmt all' to format all types of source code files we
currently are able to automatically format.

A --check flag is also added to all 'vtools fmt' commands that causes
vtools to check whether files conform to formatting guidelines and throw
an error if they don't.

Lastly, this includes a workaround for a bug in clang-format that
mistakenly identifies two header files as objective-C files.