vtools: add debuggability to cmake scripts.
By rendering the cmake commands to the file, we can inspect exactly
what is actually running on each folder including environment variables,
etc.

In addition, this sets us up to perform unit testing on vtools commands
without actually running the commands once we add a --dry-mode
to the configuration option.

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20200129223041.1262005-1-alex@vectorized.io>